Metronome: use setInitialDelay to set next click
As described in a previous commit message, the original code calculated
the "next next" click interval, because using setInterval does not
clear already queued, which is already set to be fired after
actionPerformed is run.

In addition, create class MultiIntervalTimer as a wraparound for this.
I have reason to believe this will be useful for other issues where
complex swing groups (i.e. for time signatures not in steady time)
need to be represented.