Re-implement save files (first draft)
Add Serializable interface to all relevant objects, except Sound,
which is dependent on syscalls to open the relevant Files again.

The open/save/read/write file functions need to be isolated somehow,
and everything that comes with adding and deleting staves needs to
be contained as well.

I'm just glad this is over...